requirements.txt
langchain
streamlit
sentence_transformers
pypdf
faiss-cpu
python-dotenv
google.generativeai

google-search-results
wikipedia